requirements.txt
pycodestyle==2.7.0
python-dotenv==0.19.0
types-requests==2.25.9
typing-extensions==3.10.0.2
requests==2.25.1
clint==0.5.1